pass timeout and move to pkg/module
